refactor user profile routes
Previously the implementation used #show for #profile. They're now two separate routes.

As well, get_user is refactored to simply get the user from params instead of returning the current_user if nothing is found. This way we are more explicit about where the user is coming from. As a result, it is only used for #show.